Completed user registration and login, as well as restriction system

+++ controllers/JobsController.php
@@ -3,6 +3,7 @@
 namespace app\controllers;
 use yii\data\Pagination;
 use app\models\Job;
+use yii\filters\AccessControl;
 use Yii;
 
 class JobsController extends \yii\web\Controller
@@ -34,6 +35,10 @@ class JobsController extends \yii\web\Controller
     public function actionDelete($id)
     {
         $job = Job::findOne($id);
+        if (Yii::$app->user->identity->id !== $job->user_id) {
+            return redirect('/jobs');
+        }
+        
         $job->delete();
         Yii::$app->getSession()->setFlash('success', 'Job deleted');
         return $this->redirect('/jobs');
@@ -42,6 +47,11 @@ class JobsController extends \yii\web\Controller
     public function actionEdit($id)
     {
         $job = Job::find()->where(['id'=>$id])->one();
+
+        /* Check for the owner */
+        if (Yii::$app->user->identity->id !== $job->user_id) {
+            return $this->redirect('/jobs');
+        }
 
         if ($job->load(Yii::$app->request->post())) {
             if ($job->validate()) {
@@ -71,6 +81,24 @@ class JobsController extends \yii\web\Controller
             $jobs = Job::find()->where(['id' => $id])->all();           
             return $this->render('index', ['jobs' => $jobs]);
         }
+    }
+
+    /* Restrictions */
+    public function behaviors()
+    {
+        return [
+            'access' => [
+                'class' => AccessControl::className(),
+                'only' => ['create', 'edit', 'delete'],
+                'rules' => [
+                    [
+                        'actions' => ['create', 'edit', 'delete'],
+                        'allow' => true,
+                        'roles' => ['@'],
+                    ],
+                ],
+            ]
+        ];
     }
 
 }
